Allow selecting dark/light themes
Because of a limitation in the UWP platform, the theme transition is not great. "The theme can only be set when the app is started, not while it’s running."[1] Currently, the program stores the chosen theme in the apps settings and loads it at runtime. A possible improvement would be to notify the user of the need to restart and offer the option to do that.

[1] https://docs.microsoft.com/en-us/uwp/api/Windows.UI.Xaml.Application#Windows_UI_Xaml_Application_RequestedTheme